+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Serialize Models to Data-Array for JSON-Export](https://github.com/5pm-HDH/churchtools-api/pull/103)
 - [Retrieve Birthdays](https://github.com/5pm-HDH/churchtools-api/pull/104)
 - [Retrieve Tags of Person](https://github.com/5pm-HDH/churchtools-api/pull/110)
+- [Added force flag to person creation, to add persons with same name](https://github.com/5pm-HDH/churchtools-api/pull/107)
 - [Absence-API](https://github.com/5pm-HDH/churchtools-api/pull/111)
 - [File-API](https://github.com/5pm-HDH/churchtools-api/pull/114)